requirements.txt
flask
flask-socketio
python-dotenv
requests
gunicorn
openpyxl
mysql-connector-python
vosk
ffmpeg-python
concurrent-futures; python_version < "3.2"
